phase3: split Species methods into src/species.cpp
Also fix remaining ODR violations in troll.hpp (mpi_rank, mpi_size,
easympi_rank were unconditional definitions; now extern + defined in troll.cpp).
